Move packets to own file, add methodPacket
